Make sure listeners get all updates when a server is expired
Use the StateChangedEvent.ChangeEvent.Time to figure out if
the current state of the receiver is older than the received
event.

We can't rely on evt.State.LastChanged for this because,
when expiring a server, we get multiple events coming in on
UrlListener.eventChannel and each StateChangedEvent created will
contain a snapshot of the same state. This will make the receiver
process only the first incoming event and drop the rest, because
rcvr.CurrentState.LastChanged = evt.State.LastChanged after the
first event gets processed.